refactor: use uv
