+++ .gitignore
@@ -1,4 +1,6 @@
 venv/
+.venv/
+.uv-cache/
 __pycache__/
 *.pyc
 *.pyo
