--- requirements.txt
@@ -1,8 +1,0 @@
-torch>=2.0.0
-tiktoken>=0.5.0
-numpy>=1.24.0
-tqdm>=4.65.0
-requests>=2.31.0
-mwparserfromhell>=0.6.5
-beautifulsoup4>=4.12.0
-lxml>=5.0.0
